requirements.txt
openai==0.27.2
google-cloud-bigquery
